Sucursal Service

+++ src/main/java/com/example/horarios/Repositorio/SucursalRepositorio.java
@@ -5,6 +5,6 @@ import com.example.horarios.Entidades.Sucursal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface SucursalRepositorio extends JpaRepository<Sucursal, Integer> {
+public interface SucursalRepositorio extends JpaRepository<Sucursal, Long> {
     
 }
